add new module to add event listeners
