+++ src/index.js
@@ -6,21 +6,11 @@ import toDoList from './modules/toDoList_data.js';
 import removeTask from './modules/remove_task.js';
 import modifyCurrentTask from './modules/modify_current_task.js';
 import statusHandler from "./modules/task_status.js";
+import upDateListeners from './modules/update_listeners.js';
 
 display(toDoList);
 
-function upDateListeners() {
-  const allDeleteButton = document.querySelectorAll('.delete-button');
-  const allTaskAsInput = document.querySelectorAll('.current-task-input');
-  const AllCheckBox = document.querySelectorAll('.square')
-  console.log(AllCheckBox)
-  allDeleteButton.forEach((button) => { button.addEventListener('click', removeTask); });
-  allTaskAsInput.forEach((inputElement) => { inputElement.addEventListener('keydown', modifyCurrentTask); });
-  AllCheckBox.forEach(checkBox => { checkBox.addEventListener('change', statusHandler.change ) });
-
-}
 upDateListeners();
-/* Add task */
 
 const inputForNewTask = document.getElementById('add-task');
